pyproject.toml
[project]
name = "skillengine"
dynamic = ["version"]
description = "A standalone skills execution engine for LLM agents"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Alex Zhang", email = "alex@example.com" }
]
keywords = ["llm", "agent", "skills", "ai", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "watchfiles>=0.21",
    "rich>=13.0",
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
    "pypdf>=6.6.2",
    "pillow>=12.1.0",
    "markitdown[pptx]>=0.1.4",
    "reportlab>=4.4.9",
    "partial-json-parser>=0.1",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.18"]
websockets = ["websockets>=12.0"]
memory = ["httpx>=0.24"]
web = ["starlette>=0.27", "uvicorn>=0.23"]
sandbox = ["boxlite>=0.1"]
a2a-server = ["fastapi>=0.100", "uvicorn>=0.23"]
claude-sdk = ["claude-agent-sdk>=0.1"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
    "mypy>=1.8",
    "types-PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/sawzhang/skillengine"
Repository = "https://github.com/sawzhang/skillengine"

[project.scripts]
skills = "skillengine.cli:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"

[tool.hatch.build.hooks.vcs]
version-file = "src/skillengine/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/skillengine"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = false
disallow_untyped_defs = false
disallow_untyped_calls = false

[[tool.mypy.overrides]]
module = [
    "skillengine.agent",
    "skillengine.session.*",
    "skillengine.memory.*",
    "skillengine.runtime.*",
    "skillengine.adapters.openai",
    "skillengine.extensions.manager",
    "skillengine.context",
    "skillengine.cli",
    "skillengine.transports.*",
    "skillengine.tools.*",
    "skillengine.modes.*",
    "skillengine.packages.*",
    "skillengine.utils.*",
    "skillengine.web.*",
]
disable_error_code = [
    "assignment",
    "arg-type",
    "attr-defined",
    "dict-item",
    "union-attr",
    "unused-ignore",
    "var-annotated",
    "import-not-found",
    "import-untyped",
    "override",
    "type-var",
    "misc",
    "no-untyped-def",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
